% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{avaliar}
\alias{avaliar}
\title{Avaliar}
\usage{
avaliar(dados)
}
\arguments{
\item{dados}{Banco de dados.}
}
\value{
Lista com ganho de informação, valor do corte, coluna do corte e probabilidades.
}
\description{
Avaliar
}
