% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{avaliar_arvore_Hold_Out}
\alias{avaliar_arvore_Hold_Out}
\title{Hold-Out}
\usage{
avaliar_arvore_Hold_Out(
  data,
  coluna_resposta,
  n_linhas_treino,
  g_info_min = 0.8,
  profundidade_max = 3
)
}
\arguments{
\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}

\item{n_linhas_treino}{Número de linhas do banco para serem usadas como treino.}

\item{g_info_min}{Valor mínimo para o ganho de informação.}

\item{profundidade_max}{Profundidade máxima dos ramos.}

\item{dados}{Banco de dados.}
}
\value{
Lista com matriz confusão, acuracia, recall e precisão.
}
\description{
Hold-Out
}
\examples{
 avaliar_arvore_Hold_Out(iris, "Species", 50, g_info_min = 0.05, profundidade_max = 3)

}
