% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{avaliar_arvore_K_fold}
\alias{avaliar_arvore_K_fold}
\title{K-Fold}
\usage{
avaliar_arvore_K_fold(
  data,
  coluna_resposta,
  g_info_min = 0.025,
  profundidade_max = 3,
  valor_p_chapeu = 0.5
)
}
\arguments{
\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}

\item{g_info_min}{Valor mínimo para o ganho de informação.}

\item{profundidade_max}{Profundidade máxima dos ramos.}

\item{dados}{Banco de dados.}
}
\value{
Lista com acuracia, recall e precisão.
}
\description{
K-Fold
}
\examples{
 avaliar_arvore_K_fold(iris, "Species", g_info_min = 0.05)

}
