% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_infor_gain.R
\name{calculate_info_gain}
\alias{calculate_info_gain}
\title{Calculate information gain}
\usage{
calculate_info_gain(data, left_data, right_data, target_col)
}
\arguments{
\item{data}{the original data}

\item{left_data}{the left part after splitting}

\item{right_data}{the right part after splitting}

\item{target_col}{the target feature}
}
\description{
Used to calculate the information gain after splitting a node into left part and right part, the standard
for splitting a node
}
\examples{
data <- data.frame(feature1 = runif(10),feature2 = runif(10),target = sample(c("A", "B"), 10, replace = TRUE))
split_value <- median(data$feature1)
left_data <- data[data$feature1 <= split_value, ]
right_data <- data[data$feature1 > split_value, ]
calculate_info_gain(data, left_data, right_data, 'target')
}
