% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{constroi_filho}
\alias{constroi_filho}
\title{Constroi filho}
\usage{
constroi_filho(
  dados,
  coluna_resposta,
  g_info_min = 0.5,
  profundidade_max = 3,
  contador = 0
)
}
\arguments{
\item{dados}{Banco de dados.}

\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}

\item{g_info_min}{Valor mínimo para o ganho de informação.}

\item{profundidade_max}{Profundidade máxima dos ramos.}
}
\description{
Constroi filho
}
