% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{entropia}
\alias{entropia}
\title{Entropia}
\usage{
entropia(coluna_resposta)
}
\arguments{
\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}
}
\value{
Lista com o ganho de informação na coluna resposta e as probabilidades de cada categoria (p-chapéu).
}
\description{
Entropia
}
