% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_split.R
\name{find_best_split}
\alias{find_best_split}
\title{Find the best split}
\usage{
find_best_split(data, target_col)
}
\arguments{
\item{data}{the original data}

\item{target_col}{the target feature/ the splitting attribute}
}
\description{
Used to find the best split based on the largest information gain
}
\examples{
data <- data.frame(feature1 = runif(10),feature2 = runif(10),target = sample(c("A", "B"), 10, replace = TRUE))
find_best_split(data, 'target')
}
