% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ganho_inf}
\alias{ganho_inf}
\title{Ganho de Informação}
\usage{
ganho_inf(corte, coluna_resposta, coluna_corte)
}
\arguments{
\item{corte}{Valor em que será realizado o corte.}

\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}

\item{coluna_corte}{Coluna do banco em que será realizado o corte.}
}
\value{
Ganho de informação obtido.
}
\description{
Ganho de Informação
}
