% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_entropy.R
\name{information_entropy}
\alias{information_entropy}
\title{Information entropy}
\usage{
information_entropy(labels)
}
\arguments{
\item{labels}{the target features of the given data}
}
\description{
Information entropy is used to quantify the impurity or disorder in a dataset
}
\examples{
information_entropy(data[target_columns])
}
