% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{organizar_banco}
\alias{organizar_banco}
\title{Organizar Banco}
\usage{
organizar_banco(banco, coluna_resposta)
}
\arguments{
\item{banco}{Banco de Dados.}

\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}
}
\value{
Banco de Dados com a coluna resposta sendo a primeira coluna do banco.
}
\description{
Organizar Banco
}
