% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{otimizar}
\alias{otimizar}
\title{Otimizar}
\usage{
otimizar(coluna_resposta, coluna_corte)
}
\arguments{
\item{coluna_resposta}{Coluna resposta do banco do tipo "categórica".}

\item{coluna_corte}{Coluna do banco em que será realizado o corte.}
}
\value{
Lista com o ganho de informação obtido, valor do corte e probabilidades.
}
\description{
Otimizar
}
