% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.DecisionTreeObject.R
\name{plot.DecisionTreeObject}
\alias{plot.DecisionTreeObject}
\title{plotDecisionTree}
\usage{
\method{plot}{DecisionTreeObject}(DecisionTreeObject, dict, green = 1,
  left = 1)
}
\arguments{
\item{DecisionTreeObject}{object of class DecisionTreeObject generated by function decisionTree}

\item{dict}{a data.frame of optional names for nodes, colnames of data.frame are: vName and toPlot, remember about stringsAsFactors=F}

\item{green}{which class should be coloured in green? 0 or 1?}

\item{left}{which class should be mentioned on leaves}
}
\description{
plot a decision tree
}
\examples{
d <- iris[, c("Species", "Sepal.Length", "Sepal.Width")]
d$Species <- as.character(d$Species)
d$Species[d$Species != "setosa"] <- "non-setosa"
x <- d$Sepal.Length
x[d$Sepal.Length <= 5.2] <- "Very Short"
x[d$Sepal.Length >  5.2 & d$Sepal.Length <= 6.1] <- "Short"
x[d$Sepal.Length >  6.1 & d$Sepal.Length <= 7.0] <- "Long"
x[d$Sepal.Length >  7.0] <- "Very Long"
d$Sepal.Length <- x
decTree <- decisionTree(d, eta=5, purity=0.95, minsplit=0)
plot(decTree)
}
\keyword{decision}
\keyword{plot}
\keyword{tree}
\keyword{tree,}
