% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dataset.R
\name{predict_dataset}
\alias{predict_dataset}
\title{Predict a dataset}
\usage{
predict_dataset(tree, dataset)
}
\arguments{
\item{tree}{the trained decision tree}

\item{dataset}{the new dataset}
}
\description{
Generalize the predict_point to a new dataset
}
\examples{
predict_dataset(tree_model, test_data)
}
