% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_point.R
\name{predict_point}
\alias{predict_point}
\title{Predict a point}
\usage{
predict_point(tree, data_point)
}
\arguments{
\item{tree}{the trained decision tree}

\item{data_point}{a given new data point}
}
\description{
Make a prediction of a new data point along the decision tree
}
\examples{
data <- data.frame(feature1 = runif(10),feature2 = runif(10),target = sample(c("A", "B"), 10, replace = TRUE))
tree <- myrpart(data, "target", max_depth = 3)
predict_point(tree, data[1, ])
}
