% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subtabela}
\alias{subtabela}
\title{Subtabelas}
\usage{
subtabela(dados)
}
\arguments{
\item{dados}{Banco de dados.}
}
\value{
Lista com as duas tabelas obtidas com o corte, nome da coluna em que foi feito o corte e valor do corte.
}
\description{
Subtabelas
}
