% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_point.R
\name{predict_point}
\alias{predict_point}
\title{Predict a point}
\usage{
predict_point(tree, data_point)
}
\arguments{
\item{tree}{the trained decision tree}

\item{data_point}{a given new data point}
}
\description{
Make a prediction of a new data point along the decision tree
}
\examples{
predict_point(tree, dataset[i, ])
}
